Update blockquote and arrow style
Reinstated the blockquote box-shadow effect and resized and realigned.
Changed the blockquote style for small devices to make the blockquote
more symmetrical to appear more attractive on small screens. Added a
highlighted/glowing affect to the arrow when hovered over using both
background-color and box-shadow properties, including a 200ms transition
so the transition does not appear abrupt.